+++ src/main/java/nextstep/subway/infrastructure/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.infrastructure;
 
+import nextstep.subway.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
