+++ src/main/java/nextstep/subway/infrastructure/StationRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.infrastructure;
 
+import nextstep.subway.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
